chore: extracts display widget `highlight` methods to a trait
