--- src/gui/widgets/key_event_handler.rs
@@ -1,5 +1,0 @@
-use crossterm::event::KeyEvent;
-
-pub trait KeyEventHandler {
-    fn on_input(&mut self, input: KeyEvent);
-}
